More module refactorings (just moving some stuff into separate files).
